first try on pipeline
Trying 2 stages pipeline, put pipeline register between ID and EX,
but regfile still put in the latter half.

There is no stall for branch and jump instructions, so natually,
now it has delay slot and the branch and jump instructions are not
correct. For example, beq rs1, rs2, imm should set pcnextbr=pc+imm,
but now it's pcnextbr=pc+4+imm, just like what MIPS does.

This is easy for MIPS. No hazard for branches.
But for RISC-V, need to stall branch and fix it to pc+imm