Implemented 2 stage pipeline
IF and ID are put into the first stage, the rest is the other stage.
But the regfile is put into the second stage, so right now, don't need
to worry about data hazards.

Only need to handle control hazard, those are branch and jump
instructions.

Later, need to put the hazard control logic into a module.  Now it has
stallF and flashE two signals.

Note: The pipeline control signals may be different when having more
pipeline stages.

In the fetch stage, when sees bxx or jump, immediately signal stallF
so the pc will stop right there wait for the branch instruction's
result, which are pcnextE and pcsrcE. When pcsrcE is set, the pc take
the new value pcnextE which either be pc+4 or the branch target.

When flashing a pipeline stage register, it can be considered as the
first half and the second half. In this case (Don't knonw about other
cases right now), set flashE, so in the next clock cycle, the input to
the pipeline register will be all 0. By doing this, the stuff that
already in the pipeline register will not be cleared and will move
along to the next stage.

The flashE signal is set by
wire flashE = (((branchtypeE != `CPU6_BRANCHTYPE_NOBRANCH) | jumpE));

Notice that the signals are taken from the latter stage, branchtypeE
not NOBRANCH and jumpE. Because, for example, beq, if flashing
pipeline register when sees it in IF, then beq will never be in the
second stage, so it will never be executed, pcsrcE will never be
set. (It may be different if the ID stage implemented.)

Also, an important concept in pipeline is bubble.  When the IF(pc)
stops there and the reset pipeline move forward, there will be bubbles
in the pipeline. Bubble is the all-zeros in the pipeline register. The
bubble should not have any effect on the cpu, for example, no register
write, no memory write. In this case, also no branch. So it won't
affect the flashE signal.

But CPU6_BRANCHTYPE_NOBRANCH is defined as 3'b010, and 3'b000 is
actually CPU6_BRANCHTYPE_BEQ. So I have to switch it. Now
CPU6_BRANCHTYPE_NOBRANCH is 3'b000, otherwise the flashE will be
signaled.

Pipeline bubble is all-zero due to the flashing. (Or maybe some other
form of data that is benign?)