change to altera 2-port RAM
8KB size
Missed cpu6_hazardunit.v
Timescale now is 1ns, need to set it when using modelsim.
Delete time delay in cpu6_dffs.v

Due to the systhesisable RAM, the timing is different than the
pseudo cache. When pcF update, it has a little delay due to the
pcreg dff.
So, need to use pcnextF instead, it's the right pc value, and
shortly it will be send to pcreg. But pcnextF starts at 0x4 when
pc starts at 0x0. Need to use reset signal to do the trick.