Propagation delays are important! qout_r <= #1 dnxt;
 #1 delay in cpu6_dffs.v are important. Without this, in the wave
view, everything is prefectly synced to CLK. In reality, not the
case. For example, even in simulationi, Altera RAM will take the
delay in count. When the address comes out of pipeline register,
it's got a bit delay (#1), the RAM won't take this address, it already
use the previous one.